feat(visual): algorithm visualization for Djikstra added
