+++ src/main/kotlin/controller/GraphPainterByCommunity.kt
@@ -1,10 +1,8 @@
 package controller
 
-import androidx.compose.ui.graphics.Color
 import model.graph.Graph
 import model.community.Louvain
 import viewmodel.graph.GraphViewModel
-import kotlin.random.Random
 
 class GraphPainterByCommunity(private val graph: Graph, private val graphViewModel: GraphViewModel) {
     private val finder = Louvain(graph)
@@ -18,13 +16,5 @@ class GraphPainterByCommunity(private val graph: Graph, private val graphViewMod
                 graphViewModel.verticesView[currVertex]!!.color=communityColor
             }
         }
-    }
-
-    private fun generateRandomColor(base: Int): Color {
-        val mRandom = Random(base)
-        val red: Int = (base + mRandom.nextInt(256)) / 2
-        val green: Int = (base + mRandom.nextInt(256)) / 2
-        val blue: Int = (base + mRandom.nextInt(256)) / 2
-        return Color(red, green, blue)
     }
 }
