+++ src/main/kotlin/controller/GraphPainterByKosaraju.kt
@@ -1,11 +1,8 @@
 package controller
 
-import androidx.compose.ui.graphics.Color
 import model.algorithms.Kosaraju
 import model.graph.Graph
-import model.community.Louvain
 import viewmodel.graph.GraphViewModel
-import kotlin.random.Random
 
 class GraphPainterByKosaraju(private val graph: Graph, private val graphViewModel: GraphViewModel) {
     private val finder = Kosaraju(graph)
@@ -19,13 +16,5 @@ class GraphPainterByKosaraju(private val graph: Graph, private val graphViewMode
                 graphViewModel.verticesView[currVertex]!!.color=communityColor
             }
         }
-    }
-
-    private fun generateRandomColor(base: Int): Color {
-        val mRandom = Random(base)
-        val red: Int = (base + mRandom.nextInt(256)) / 2
-        val green: Int = (base + mRandom.nextInt(256)) / 2
-        val blue: Int = (base + mRandom.nextInt(256)) / 2
-        return Color(red, green, blue)
     }
 }
